apps/api/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
openai
mcp
